Add rollback methods to interactors

+++ app/interactors/comments/save.rb
@@ -2,14 +2,14 @@ module Comments
   class Save
     include Interactor
 
-    delegate :comment, to: :context
+    delegate :comment, :action, to: :context
 
     def call
       comment.save || context.fail!
     end
 
     def rollback
-      comment.destroy
+      comment.destroy if action == "create"
     end
   end
 end
